+++ wkdan_Java/src/Main/java/BOJ_03/p10951.java
@@ -1,21 +1,20 @@
-package BJ_03;
+package BOJ_03;
 
 import java.io.*;
 import java.util.StringTokenizer;
 
-public class p10950 {
+public class p10951 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
-        int T = Integer.parseInt(br.readLine());
+        String input = "";
 
-        for (int i = 0; i < T; i++) {
-            StringTokenizer st = new StringTokenizer(br.readLine());
-            int B = Integer.parseInt(st.nextToken());
-            int A = Integer.parseInt(st.nextToken());
-
-            bw.write(A+B+"\n");
+        while ((input = br.readLine()) != null && !input.isEmpty()) {
+            StringTokenizer st = new StringTokenizer(input);
+            int x = Integer.parseInt(st.nextToken());
+            int y = Integer.parseInt(st.nextToken());
+            bw.write((x + y) +"\n");
         }
         bw.flush();
         bw.close();
