+++ wkdan_Java/src/Main/java/BOJ_03/p11021.java
@@ -1,22 +1,22 @@
-package BJ_03;
+package BOJ_03;
 
 import java.io.*;
 import java.util.StringTokenizer;
 
-public class p15552 {
+public class p11021 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         int t = Integer.parseInt(br.readLine());
 
-        for (int i = 0; i < t; i++) {
+        for (int i = 1; i <= t; i++) {
             StringTokenizer st = new StringTokenizer(br.readLine());
 
             int a = Integer.parseInt(st.nextToken());
             int b = Integer.parseInt(st.nextToken());
 
-            bw.write(a+b+"\n");
+            bw.write("Case #"+i+": "+ (a+b) +"\n");
         }
         bw.flush();
         bw.close();
