+++ wkdan_Java/src/Main/java/BOJ_03/p15552.java
@@ -1,24 +1,22 @@
-package BJ_03;
+package BOJ_03;
 
 import java.io.*;
 import java.util.StringTokenizer;
 
-public class p10952 {
+public class p15552 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
+        int t = Integer.parseInt(br.readLine());
 
-        for (int i = 1;; i++) {
+        for (int i = 0; i < t; i++) {
             StringTokenizer st = new StringTokenizer(br.readLine());
-            int x = Integer.parseInt(st.nextToken());
-            int y = Integer.parseInt(st.nextToken());
-            if (x != 0 && y != 0) {
-                bw.write(x + y + "\n");
-            } else {
-                break;
 
-            }
+            int a = Integer.parseInt(st.nextToken());
+            int b = Integer.parseInt(st.nextToken());
+
+            bw.write(a+b+"\n");
         }
         bw.flush();
         bw.close();
