+++ wkdan_Java/src/Main/java/BOJ_03/p2438.java
@@ -1,9 +1,8 @@
-package BJ_03;
+package BOJ_03;
 
 import java.io.*;
-import java.util.StringTokenizer;
 
-public class p2439 {
+public class p2438 {
     public static void main(String[] args ) throws IOException {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -11,7 +10,7 @@ public class p2439 {
         int n = Integer.parseInt(br.readLine());
 
         for (int i = 1; i <= n; i ++) {
-            bw.write((" ").repeat(n-i)+("*").repeat(i)+"\n");
+            bw.write(("*").repeat(i)+"\n");
 
         }
         bw.flush();
