+++ wkdan_Java/src/Main/java/BOJ_03/p2739.java
@@ -1,4 +1,4 @@
-package BJ_03;
+package BOJ_03;
 
 import java.io.*;
